+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/recyclerview/ConferenceRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package com.emmsale.presentation.ui.main.event.conference
+package com.emmsale.presentation.ui.main.event.conference.recyclerview
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
